# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Documentation",
    "version": "1.0"
  },
  "paths": {
...
    "/admin/category/page": {
      "get": {
        "summary": "分页查询分类",
        "description": "分页查询分类",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": false
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": false
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "分类名称",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": false
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "分类类型 1菜品分类  2套餐分类",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "编码：1成功，0和其它数字为失败"
                    },
                    "msg": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "description": "总记录数",
                          "format": "int64"
                        },
                        "records": {
                          "type": "array",
                          "description": "当前页数据集合",
                          "items": {
                            "type": "object",
                            "properties": {},
                            "description": "java.util.List"
                          }
                        }
                      },
                      "description": "数据"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "管理员端"
        ]
      }
    },
    "/admin/category": {
      "put": {
        "summary": "修改分类",
        "description": "修改分类",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "主键",
                    "format": "int64"
                  },
                  "type": {
                    "type": "integer",
                    "description": "类型 1 菜品分类 2 套餐分类"
                  },
                  "name": {
                    "type": "string",
                    "description": "分类名称"
                  },
                  "sort": {
                    "type": "integer",
                    "description": "排序"
                  }
                },
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "编码：1成功，0和其它数字为失败"
                    },
                    "msg": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "description": "数据"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "管理员端"
        ]
      },
      "post": {
        "summary": "新增菜品分类",
        "description": "新增菜品分类",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "主键",
                    "format": "int64"
                  },
                  "type": {
                    "type": "integer",
                    "description": "类型 1 菜品分类 2 套餐分类"
                  },
                  "name": {
                    "type": "string",
                    "description": "分类名称"
                  },
                  "sort": {
                    "type": "integer",
                    "description": "排序"
                  }
                },
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "编码：1成功，0和其它数字为失败"
                    },
                    "msg": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "description": "数据"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "管理员端"
        ]
      },
      "delete": {
        "summary": "删除分类",
        "description": "删除分类",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "编码：1成功，0和其它数字为失败"
                    },
                    "msg": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "description": "数据"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "管理员端"
        ]
      }
    },
    "/admin/category/status/{status}": {
      "post": {
        "summary": "禁用或启用分类",
        "description": "禁用或启用分类",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "编码：1成功，0和其它数字为失败"
                    },
                    "msg": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "description": "数据"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "管理员端"
        ]
      }
    },
    "/admin/category/list": {
      "get": {
        "summary": "查询分类列表",
        "description": "查询分类列表",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "编码：1成功，0和其它数字为失败"
                    },
                    "msg": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "data": {
                      "type": "array",
                      "description": "数据",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "主键",
                            "format": "int64"
                          },
                          "type": {
                            "type": "integer",
                            "description": "类型 1 菜品分类 2 套餐分类"
                          },
                          "name": {
                            "type": "string",
                            "description": "分类名称"
                          },
                          "sort": {
                            "type": "integer",
                            "description": "排序"
                          }
                        },
                        "description": "com.zhiwei.dto.CategoryDTO"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "管理员端"
        ]
      }
    },
...
    "/user/category/list": {
      "get": {
        "summary": "查询分类",
        "description": "查询分类",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "编码：1成功，0和其它数字为失败"
                    },
                    "msg": {
                      "type": "string",
                      "description": "错误信息"
                    },
                    "data": {
                      "type": "array",
                      "description": "数据",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "",
                            "format": "int64"
                          },
                          "type": {
                            "type": "integer",
                            "description": "类型: 1菜品分类 2套餐分类"
                          },
                          "name": {
                            "type": "string",
                            "description": "分类名称"
                          },
                          "sort": {
                            "type": "integer",
                            "description": "顺序"
                          },
                          "status": {
                            "type": "integer",
                            "description": "分类状态 0标识禁用 1表示启用"
                          },
                          "createTime": {
                            "type": "string",
                            "description": "创建时间"
                          },
                          "updateTime": {
                            "type": "string",
                            "description": "更新时间"
                          },
                          "createUser": {
                            "type": "integer",
                            "description": "创建人",
                            "format": "int64"
                          },
                          "updateUser": {
                            "type": "integer",
                            "description": "修改人",
                            "format": "int64"
                          }
                        },
                        "description": "com.zhiwei.entity.Category"
                      }
                    }
                  },
                  "description": ""
                }
              }
            }
          }
        },
        "tags": [
          "用户C端"
        ]
      }
    },
...
  },
...
}